requirements.txt
wheel==0.34.2
SQLAlchemy==1.3.16
SQLAlchemy-Utils==0.36.5
urllib3==1.25.9
URLObject==2.4.3
uWSGI==2.0.18
alembic==1.4.2
arrow==0.15.6
docutils==0.16
flasgger==0.9.4
Flask==1.1.2
Flask-Cors==3.0.8
Flask-Dance==3.0.0
Flask-JWT-Extended==3.24.1
Flask-Migrate==2.5.3
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
Flask-Testing==0.8.0
Jinja2==2.11.2
jmespath==0.9.5
jsonschema==3.2.0
jsontools==0.1
lazy==1.4
Mako==1.1.2
MarkupSafe==1.1.1
pycparser==2.20
Pygments==2.6.1
PyJWT==1.7.1
PyMySQL==0.9.3
PyNaCl==1.3.0
pyOpenSSL==19.1.0
pyparsing==2.4.7
pystache==0.5.4
python-dateutil==2.8.1
python-editor==1.0.4
python-http-client==3.2.7
pytz==2020.1
PyYAML==5.3.1
qrcode==6.1
oauthlib==3.1.0